optimize webpack configurations
