+++ index.html
@@ -5,8 +5,8 @@
 	<meta charset="utf-8">
 	<meta name="viewport" content="width=device-width,initial-scale=1">
 	<title>認知特性テスト</title>
-	<link rel="stylesheet" type="text/css" href="src/css/normalize.css">
-	<link rel="stylesheet" type="text/css" href="src/css/base.css">
+	<link rel="stylesheet" type="text/css" href="src/normalize.css">
+	<link rel="stylesheet" type="text/css" href="src/base.css">
 	<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 </head>
