+++ package.json
@@ -2,14 +2,14 @@
   "name": "personal-cognitive-property-test",
   "version": "1.0.0",
   "description": "Web application for personal cognitive property test",
-  "main": "src/js/app.js",
+  "main": "src/index.js",
   "repository": "git@github.com:yutarot/personal-cognitive-property-test.git",
   "author": "yutarot <takenoco88@gmail.com>",
   "license": "MIT",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build:production": "cross-env NODE_ENV=production webpack",
-    "build:development": "webpack",
+    "build": "webpack --config webpack.prod.js",
+    "dev": "webpack --config webpack.dev.js",
     "start": "webpack --watch -d & browser-sync start --server --files='./*.html, ./dist/*'"
   },
   "devDependencies": {
@@ -17,9 +17,12 @@
     "babel-loader": "^7.1.4",
     "babel-preset-es2015": "^6.24.1",
     "browser-sync": "^2.23.6",
+    "clean-webpack-plugin": "^0.1.19",
     "cross-env": "^5.1.4",
+    "uglifyjs-webpack-plugin": "^1.2.4",
     "webpack": "^4.2.0",
-    "webpack-cli": "^2.0.12"
+    "webpack-cli": "^2.0.12",
+    "webpack-merge": "^4.1.2"
   },
   "dependencies": {
     "chart.js": "^2.7.2",
