+++ webpack.prod.js
@@ -1,0 +1,12 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+
+module.exports = merge(common, {
+    mode: 'production',
+    plugins: [
+        new UglifyJSPlugin({
+            sourceMap: true,
+        })
+    ]
+});
